requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python_multipart


